Added user review history view. Minor cosmetic changes in Login, Register and AddBook views

+++ BookRecommendationWebApp/BookRecommendationWebApp/Models/BookDetailsViewModel.cs
@@ -15,7 +15,7 @@ namespace BookRecommendationWebApp.Models
         public string ImageFileName { get; set; }
         public List<Category> Categories { get; set; }
         public int UserRating { get; set; }
-        public float AverageRating { get; set; }
+        public double AverageRating { get; set; }
         public int RatingsCount { get; set; }
     }
 }
